requirements.txt
# requirements.txt for QuasiHedge (Python 3.11)
# Core numeric / data
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.10.1

# Machine learning
scikit-learn>=1.2.2
lightgbm>=3.3.5
joblib>=1.3.1

# Embeddings / transformers (sentence-transformers requires torch)
sentence-transformers>=2.2.2
transformers>=4.30.0
# NOTE: choose torch binary for your platform (see notes below)
torch>=2.1.0

# NLP
nltk>=3.8.1
spacy>=3.5.0

# Optional nicer English model (download separately, not via pip)
# spacy model command: python -m spacy download en_core_web_sm

# Web / API clients
yfinance>=0.2.21
alpha_vantage>=2.3.1
pycoingecko>=1.0.3
feedparser>=6.0.10
requests>=2.31.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Telegram / messaging
python-telegram-bot>=20.4
telethon>=1.30.0

# DB / persistence / migrations
SQLAlchemy>=2.0.18
alembic>=1.11.1
psycopg2-binary>=2.9.7

# Monitoring & Sentry
prometheus-client>=0.17.0
sentry-sdk>=1.17.0

# Explainability / plotting / utilities
shap>=0.42.1
matplotlib>=3.7.2
tqdm>=4.65.0

# Caching / optional redis
redis>=4.5.5

# Development / testing
python-dotenv>=1.0.0
pytest>=7.3.2